Move enhanceCards and decideContainerPalette to their own files
